Revert the epoch bump in previous commit
Usually it is not safe to revert epoch bumps, but in this case it should
be. The epoch-bumped control-center build in rawhide (1fc7b6b3,
Christoph Wickert <cwickert@fedoraproject.org>) was uninstallable for
most people because of dependency issues with gnome-menus. Because of
that, it is unlikely that anyone actually managed to install the build.

Error: Package: 2:control-center-3.1.4-1.fc17.i686 (rawhide)
           Requires: libgnome-menu.so.2
           Removing: gnome-menus-3.0.1-1.fc16.i686 (@rawhide/16)
               libgnome-menu.so.2
           Updated By: gnome-menus-3.1.5-3.fc17.i686 (rawhide)
               Not found
Error: Package: 1:control-center-devel-3.1.3-1.fc16.i686 (@rawhide/16)
           Requires: control-center = 1:3.1.3-1.fc16
           Removing: 1:control-center-3.1.3-1.fc16.i686 (@rawhide/16)
               control-center = 1:3.1.3-1.fc16
           Updated By: 2:control-center-3.1.4-1.fc17.i686 (rawhide)
               control-center = 2:3.1.4-1.fc17